# docs/swagger.yaml
basePath: /api/v1
definitions:
...
  utils.APIResponse:
    properties:
      data: {}
      error:
        type: string
      message:
        type: string
      success:
        type: boolean
    type: object
host: localhost:8080
info:
  contact:
    email: mayank@example.com
    name: Mayank
  description: A role-based finance dashboard backend built with Go and Gin.
  license:
    name: MIT
  title: Finance Dashboard API
  version: "1.0"
paths:
  /analytics/by-category:
    get:
      description: Returns income and expense totals grouped by category, ordered
        by total descending.
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/utils.APIResponse'
        "403":
          description: Forbidden
          schema:
            $ref: '#/definitions/utils.APIResponse'
      security:
      - BearerAuth: []
      summary: Get breakdown by category
      tags:
      - analytics
  /analytics/summary:
    get:
      description: Returns total income, total expenses, and net balance.
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/utils.APIResponse'
        "403":
          description: Forbidden
          schema:
            $ref: '#/definitions/utils.APIResponse'
      security:
      - BearerAuth: []
      summary: Get financial summary
      tags:
      - analytics
  /analytics/trend:
    get:
      description: Returns month-by-month income and expense totals for the last N
        months (default 6, max 24).
      parameters:
      - description: Number of months to include (default 6, max 24)
        in: query
        name: months
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/utils.APIResponse'
        "403":
          description: Forbidden
          schema:
            $ref: '#/definitions/utils.APIResponse'
      security:
      - BearerAuth: []
      summary: Get monthly income vs expense trend
      tags:
      - analytics
...
securityDefinitions:
  BearerAuth:
    description: 'Enter your JWT token as: Bearer <token>'
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"